+++ src/main/java/coupon/application/coupon/MemberCouponMapperImpl.java
@@ -5,9 +5,9 @@ import java.util.Map;
 import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-import coupon.application.member.IssuedCouponResponse;
-import coupon.application.member.MemberCouponMapper;
-import coupon.application.member.MemberCouponResponse;
+import coupon.application.membercoupon.IssuedCouponResponse;
+import coupon.application.membercoupon.MemberCouponMapper;
+import coupon.application.membercoupon.MemberCouponResponse;
 import coupon.domain.Coupon;
 import coupon.domain.CouponDiscountApply;
 import coupon.domain.CouponIssuableDuration;
