+++ src/main/java/coupon/application/membercoupon/MemberCouponResponse.java
@@ -1,4 +1,4 @@
-package coupon.application.member;
+package coupon.application.membercoupon;
 
 import java.time.LocalDateTime;
